All final chnages to refactor the code - Changed Implemenation of JoyrideTour component to manage user tours within the application. - Created and Moved code Related to Modal opening of JoyrideTour to TourModal and TourTriggerButton components for tour initiation and display. - Developed step management and click handling logic for tour navigation. - Configured styles and step mappings for different tour types. - Integrated tour functionality into TheLayout component.
