+++ frontend/lost/src/components/JoyrideTour/config/stepMaps.js
@@ -1,0 +1,53 @@
+export const pipelineStepMap = {
+  'dropdown-open': 6,
+  'datasource-selected': 7,
+  'path-selected': 9,
+  'done-clicked': 10,
+  'anno-name': 11,
+  'anno-task-info-done': 15,
+  'user-selection-done': 17,
+  'label-tree-selection-done': 19,
+  'label-selection-done': 21,
+  'storage-settings-done': 23,
+  'configuration-done': 25,
+  'template-next': 26,
+  'latest-running-pipeline': 29,
+  'latest-running-annotask': 30,
+};
+
+export const instructionStepMap = {
+  'start-instruction': 0,
+  'add-step': 1,
+  'open-file-browser': 6,
+  'save-step': 7,
+  'instruction-list': 8,
+  'edit-step': 11,
+  'pipelines-nav': 13,
+  'dropdown-open': 19,
+  'datasource-selected': 20,
+  'path-selected': 21,
+  'done-clicked': 23,
+  'anno-name': 24,
+  'anno-task-info-done': 28,
+  'user-selection-done': 30,
+  'label-tree-selection-done': 32,
+  'label-selection-done': 34,
+  'storage-settings-done': 36,
+  'configuration-done': 38,
+  'template-next': 39,
+  'anno-task-display': 42,
+};
+
+export const instructionStepMap2 = {
+  'start-instruction': 0,
+  'add-step': 1,
+  'open-file-browser': 6,
+  'save-step': 7,
+  'instruction-list': 8,
+  'edit-step': 11,
+  'pipelines-nav': 13,
+};
+
+export const labelStepMap = {
+  'latest-label-tree': 4,
+};
