+++ frontend/lost/src/coreui_containers/TheLayout.jsx
@@ -10,7 +10,7 @@ import TheContent from './TheContent'
 import TheFooter from './TheFooter'
 import TheHeader from './TheHeader'
 import TheSidebar from './TheSidebar'
-import JoyrideTour from '../components/JoyrideTour'; 
+import JoyrideTour from '../components/JoyrideTour/JoyrideTour'; 
 
 const TheLayout = () => {
     const location = useLocation();
